Rename task to interactor; switch to Rx2 fully
