+++ app/src/main/java/co/touchlab/droidconandroid/shared/network/SponsorsRequest.java
@@ -1,7 +1,7 @@
 package co.touchlab.droidconandroid.shared.network;
 
 
-import retrofit2.Call;
+import io.reactivex.Observable;
 import retrofit2.http.GET;
 import retrofit2.http.Path;
